+++ src/main/java/com/projectileoverride/ProjectileIds.java
@@ -7,21 +7,26 @@ public final class ProjectileIds
 	public static final int NONE = -1;
 
 	// All ids are Mage, Ranged, Melee.
-	public static final int[] CERBERUS = new int[] {1242, 1245, NONE};
+    public static final int[] AKKHA = new int[] {2253, 2255, NONE};
+    public static final int[] CERBERUS = new int[] {1242, 1245, NONE};
 	public static final int[] DAGGANOTH_KINGS = new int[] {162, 475, NONE};
 	public static final int[] DOOM_OF_MOKHAIOTL = new int[] {3379, 3380, 3378};
 	public static final int[] DOOM_OF_MOKHAIOTL_ROCKS = new int[] {3385, 3384, NONE};
 	public static final int[] HUEYCOATL = new int[] {2975, 2972, 2969};
 	public static final int[] HUNLLEF_NORMAL = new int[] {1707, 1711, NONE};
 	public static final int[] HUNLLEF_CORRUPTED = new int[] {1708, 1712, NONE};
+    public static final int[] HYDRA = new int[] {1662, 1663, NONE};
 	public static final int[] INFERNO = new int[] {1380, 1378, NONE};
+    public static final int[] KALPHITE_QUEEN = new int[] {280, 288, NONE};
+    public static final int[] KREE_ARRA = new int[] {1200, 1199, NONE};
 	public static final int[] LEVIATHAN = new int[] {2489, 2487, 2488};
 	public static final int[] OLM = new int[] {1341, 1343, 1345};
 	public static final int[] SCRURRIUS = new int[] {2640, 2642, NONE};
 	public static final int[] SOTETSEG = new int[] {1606, 1607, NONE};
 	public static final int[] VARDORVIS = new int[] {2520, 2521, NONE};
 	public static final int[] WARDENS = new int[] {2224, 2241, NONE};
-	public static final int[] WHISPERER = new int[] {2445, 2444, NONE};
+    public static final int[] WARDENS_DIVINE = new int[] {2208, 2206, 2204};
+    public static final int[] WHISPERER = new int[] {2445, 2444, NONE};
 	public static final int[] ZEBAK = new int[] {2181, 2187, NONE};
 	public static final int[] ZEBAK_ROCKS = new int[] {2176, 2178, NONE};
 	public static final int[] ZULRAH = new int[] {1046, 1044, NONE};